OlegMolchanov's_Py_lesson/Parsing/Avito_tel/requirements.txt
Pillow==6.2.1
pytesseract==0.3.1
selenium==3.141.0
urllib3==1.25.7
